fix(windows): lock down GUI/Tunnel pipe DACLs
Pull the unsafe Windows security-descriptor wrappers from #13153 into a
new `windows-security` workspace crate, then use it to replace the NULL
DACL on every Firezone named pipe with one that grants only `LocalSystem`,
`BUILTIN\Administrators`, and `BUILTIN\Users` (read/write/sync).

Until now the Tunnel and GUI pipes accepted connections from any local
principal — including Low-IL sandboxed processes, `NETWORK SERVICE`,
and anonymous logons — because `SetSecurityDescriptorDacl(psd, true,
None, false)` installs a NULL DACL. The new SDDL keeps the non-admin
GUI working (the user is in `BU`) while excluding service identities
and other unintended callers.

The new crate is `cfg(windows)`-only at the source level so it builds
to an empty rlib on Linux/macOS, leaving cross-platform CI green.
