+++ rust/gui-client/src-tauri/Cargo.toml
@@ -88,6 +88,7 @@ tracing-journald = { workspace = true }
 admx-macro = { workspace = true }
 tauri-winrt-notification = "0.7.2"
 windows-native-keyring-store = { workspace = true }
+windows-security = { workspace = true }
 windows-service = { workspace = true }
 winreg = { workspace = true }
 
@@ -95,8 +96,11 @@ winreg = { workspace = true }
 workspace = true
 features = [
     "Win32_Foundation",
+    "Win32_Security",
+    # For `SECURITY_ATTRIBUTES` on the pipe
+    "Win32_System_Pipes",
+    # For IPC system
     "Win32_System_Threading",
-    "Win32_System_Pipes", # For IPC system
 ]
 
 [build-dependencies]
