+++ rust/libs/windows-security/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "windows-security"
+version = "0.1.0"
+edition = { workspace = true }
+license = { workspace = true }
+description = "Safe Rust wrappers around the Windows security descriptor APIs."
+
+[dependencies]
+anyhow = { workspace = true }
+
+[target.'cfg(windows)'.dependencies.windows]
+workspace = true
+features = ["Win32_Foundation", "Win32_Security", "Win32_Security_Authorization"]
+
+[target.'cfg(windows)'.dev-dependencies]
+tempfile = { workspace = true }
+
+[lints]
+workspace = true
